czech point
