+++ editor/src/components/canvas/design-panel-root.tsx
@@ -37,7 +37,7 @@ import { CanvasToolbar } from '../editor/canvas-toolbar'
 import { useDispatch } from '../editor/store/dispatch-context'
 import { LeftPaneComponent } from '../navigator/left-pane'
 import { FloatingPanelsContainer } from './floating-panels'
-import type { Menu, Pane } from './floating-panels-state'
+import type { Menu, Pane } from './floating-panels-state-2'
 import type { ResizableProps } from '../../uuiui-deps'
 import type { Direction } from 're-resizable/lib/resizer'
 import { isFeatureEnabled } from '../../utils/feature-switches'
