+++ editor/src/components/navigator/left-pane/index.tsx
@@ -29,7 +29,7 @@ import { SettingsPane } from './settings-pane'
 import { NavigatorComponent } from '../navigator'
 import { usePubSubAtom } from '../../../core/shared/atom-with-pub-sub'
 import type { ResizeCallback } from 're-resizable'
-import type { Menu, Pane } from '../../canvas/floating-panels-state'
+import type { Menu, Pane } from '../../canvas/floating-panels-state-2'
 import type { Direction } from 're-resizable/lib/resizer'
 import { isFeatureEnabled } from '../../../utils/feature-switches'
 import { when } from '../../../utils/react-conditionals'
